fix:Fix the issue where the selected group does not match the expected one when the priority list tuple is empty



# message auto-generated for no-merge-commit merge:
!1183 merge fixbug into main

fix:Fix the issue where the selected group does not match the expected one when the priority list tuple is empty

Created-by: balabala-123
Commit-by: balabala-123
Merged-by: liwei3013
Description: <!--  感谢您提交 Pull Request！以下是一些提示：

1) 如果您是第一次提交，请阅读我们的贡献者指南：https://gitcode.com/openHiTLS/openhitls/wiki/FAQ-社区需求贡献流程.md

-->


**此 PR 的作用 / 为什么需要它**：


**自检清单**：

- [ ] **测试**：代码是否已有UT/ST测试用例覆盖，新增测试用例是否随本PR一并上库
- [ ] **接口**：是否涉及对外接口变更，变更已通过接口评审
- [ ] **文档**：是否涉及文档修改

<!-- **给审查者的特别说明**： -->
<!-- + - [ ] 是否导致无法前向兼容 -->
<!-- + - [ ] 是否涉及依赖的三方库变更 -->


See merge request: openHiTLS/openhitls!1183

+++ tls/config/src/config_group.c
@@ -264,7 +264,7 @@ int32_t ConfigLoadGroupInfo(HITLS_Config *config)
     }
     // Load default group configuration
 #ifdef HITLS_TLS_CONFIG_CIPHER_SUITE
-    return HITLS_CFG_SetGroupList(config, DEFAULT_GROUP_ID, strlen(DEFAULT_GROUP_ID));
+    return HITLS_CFG_SetGroupList(config, DEFAULT_GROUP_ID, (uint32_t)strlen(DEFAULT_GROUP_ID));
 #else
     return HITLS_CFG_SetGroup(config, DEFAULT_GROUP_ID, sizeof(DEFAULT_GROUP_ID) / sizeof(DEFAULT_GROUP_ID[0]));
 #endif
@@ -372,7 +372,7 @@ int32_t ConfigLoadGroupInfo(HITLS_Config *config)
     }
     // Load default group configuration
 #ifdef HITLS_TLS_CONFIG_CIPHER_SUITE
-    return HITLS_CFG_SetGroupList(config, DEFAULT_GROUP_ID, strlen(DEFAULT_GROUP_ID));
+    return HITLS_CFG_SetGroupList(config, DEFAULT_GROUP_ID, (uint32_t)strlen(DEFAULT_GROUP_ID));
 #else
     return HITLS_CFG_SetGroup(config, DEFAULT_GROUP_ID, sizeof(DEFAULT_GROUP_ID) / sizeof(DEFAULT_GROUP_ID[0]));
 #endif
